refactor(ui): revisit ui handlers
